+++ lib/obsobject.dart
@@ -1,4 +1,3 @@
-export 'src/cancelablethen.dart';
 export 'src/commission.dart';
 export 'src/computed.dart';
 export 'src/observable.dart';