+++ lib/src/observablebase.dart
@@ -1,4 +1,6 @@
 import 'dart:async';
+
+import 'computed.dart';
 import 'validator.dart';
 
 ///A listener on an observable, it can close listen in later
@@ -22,7 +24,12 @@ class Subscription {
 abstract class ObservableBase<T> {
   final _callbacks = <Function>[];
   StreamController? _streamer;
-  ObservableValidator? _isValid;
+  final Validator? _validator;
+  Computed<String?>? _error;
+  Computed<bool>? _valid;
+
+  ///Create an observable with validator
+  ObservableBase(this._validator);
 
   ///The old value if has change
   T get oldValue;
@@ -102,151 +109,41 @@ abstract class ObservableBase<T> {
   }
 
   ///Check there are listeners on this observable
-  bool get hasListener =>
-      (_isValid == null && _callbacks.isNotEmpty) ||
-      (_isValid != null && (_isValid!.hasListener || _callbacks.length > 1));
+  bool get hasListener => _callbacks.isNotEmpty;
 
   ///Check value has updated
   bool get modified => peek != oldValue;
 
-  ///An observable keep current status validation
-  ObservableValidator get isValid {
-    _isValid ??= ObservableValidator(this);
-    return _isValid!;
+  ///Check has using validate value
+  bool get hasValidator => _validator != null;
+
+  ///An error message if observable value is invalid
+  Computed<String?> get error {
+    _error ??= Computed<String?>(() {
+      if (!hasValidator) return null;
+      return _validator!.validate(value);
+    });
+    return _error!;
   }
 
-  ///Check has using validate value
-  bool get hasValidator => _isValid != null;
+  ///Valid state of observable
+  Computed<bool> get valid {
+    _valid ??= Computed<bool>(() => error.value == null);
+    return _valid!;
+  }
 
   ///Close stream (if used) and all listeners on this observable
   void dispose() {
     _streamer?.close();
     _streamer = null;
-    _isValid?.dispose();
     _callbacks.clear();
+    _error?.dispose();
+    _valid?.dispose();
   }
 
   @override
   String toString() {
     return '${runtimeType.toString()}(${peek.toString()})';
-  }
-}
-
-///Handle valid state for an observable
-class ObservableValidator extends ObservableBase<bool> {
-  Validator? _validator;
-
-  ///Current observable use validate
-  final ObservableBase observable;
-  bool _oldValue = true;
-  bool _value = true;
-  String _oldMessage = '';
-  String _message = '';
-  bool _hasChanged = true;
-  dynamic _oldObsValue;
-
-  ///Create a observable validation
-  ObservableValidator(this.observable) : super() {
-    //insert at top list, so it can run update valid status
-    //before other listeners called
-    observable._callbacks.insert(0, _observableChanged);
-  }
-
-  void _observableChanged() {
-    _hasChanged = true;
-    if (hasListener) _update();
-  }
-
-  @override
-  bool get oldValue => _oldValue;
-
-  @override
-  bool get peek {
-    //check computed value
-    if (!_hasChanged && _oldObsValue != observable.peek) _hasChanged = true;
-    _update();
-    return _value;
-  }
-
-  ///Get current validator used to validate observable value
-  Validator? get validator => _validator;
-
-  ///Set current validator used to validate observable value
-  ///It also force update, notify current status validation
-  set validator(Validator? val) {
-    if (_validator == val) return;
-    _validator = val;
-    _hasChanged = true;
-    if (hasListener) _update();
-    _notifyObservable();
-  }
-
-  ///A current invalid message, it doesn't has null, just empty or not
-  String get message {
-    //check computed value
-    if (!_hasChanged && _oldObsValue != observable.peek) _hasChanged = true;
-    _update();
-    return _message;
-  }
-
-  ///update current valid status for observable
-  ///it can still invalid but with difference message
-  void _update() {
-    if (!_hasChanged || _validator == null) return;
-    _oldObsValue = observable.peek;
-    _hasChanged = false;
-    late String? s;
-    runZoned(() {
-      s = _validator!.validate(observable.peek) ?? '';
-    }, zoneValues: {'validatorCallback': _validatorCallback});
-    if (s != _message) {
-      _oldMessage = _message;
-      _message = s ?? '';
-      _oldValue = _value;
-      _value = _message.isEmpty;
-      notify();
-    }
-  }
-
-  ///Function callback by ValidatorAsync, they're communicate by Zone values
-  void _validatorCallback(String? msg) {
-    //wait for other task changed finish
-    //print('Validator callback: $msg');
-    var s = msg ?? '';
-    if (s != _message) {
-      _oldMessage = _message;
-      _message = s;
-      _oldValue = _value;
-      _value = _message.isEmpty;
-      notify();
-      _notifyObservable();
-    }
-  }
-
-  @override
-  bool get modified => _oldValue != _value || _oldMessage != _message;
-
-  @override
-  Subscription listen(Function callback) {
-    _update();
-    return super.listen(callback);
-  }
-
-  ///force all listener in observable
-  ///there are change in isValid by async validate,
-  ///due to it can listen in observable instead isValid
-  void _notifyObservable() {
-    for (var p in observable._callbacks) {
-      if (p != _observableChanged) {
-        _executeCallback(p);
-      }
-    }
-  }
-
-  @override
-  void dispose() {
-    _callbacks.remove(_observableChanged);
-    super.dispose();
   }
 }