+++ lib/src/validator.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
 
-import 'cancelablethen.dart';
-
 ///Interface to register a subclass [Validator]
 ///to determine how to create it from a [Map] data
 class ValidatorRegister {
@@ -611,7 +609,7 @@ class ValidatorAsync extends Validator {
 
   ///A debounce time (millisecond) to run validate
   int rateLimit = 0;
-  CancelableThen? _subValid;
+  Timer? _subValid;
 
   ///Create [Validator] can check value by an async function
   ValidatorAsync([this.valid]);
@@ -636,24 +634,22 @@ class ValidatorAsync extends Validator {
       _oldValue = value;
       if (onStart != null) onStart!();
       _subValid?.cancel();
-      _subValid = CancelableThen(
-          future: Future.delayed(Duration(milliseconds: rateLimit)),
-          then: (_) {
-            valid!(value).then((v) {
-              _running = false;
-              _result = v;
-              //notify manual
-              if (onDone != null) onDone!(_result);
-              //callback by Zone
-              var callback = Zone.current['validatorCallback'];
-              if (callback != null) {
-                var s = _result
-                    ? null
-                    : Validator.getMessage([message, defaultMessage]);
-                callback(s);
-              }
-            });
-          });
+      _subValid = Timer(Duration(milliseconds: rateLimit), () {
+        valid!(value).then((v) {
+          _running = false;
+          _result = v;
+          //notify manual
+          if (onDone != null) onDone!(_result);
+          //callback by Zone
+          var callback = Zone.current['validatorCallback'];
+          if (callback != null) {
+            var s = _result
+                ? null
+                : Validator.getMessage([message, defaultMessage]);
+            callback(s);
+          }
+        });
+      });
 
       return Validator.getMessage([messageAsync, defaultMessageAsync]);
     } else if (_running) {